+++ Engine/IObject.cpp
@@ -2,10 +2,10 @@
 #include "Engine/Point.hpp"
 
 namespace Engine {
-    IObject::IObject(float x, float y, float w, float h, float anchorX, float anchorY) : Position(Point(x, y)), Size(Point(w, h)), Anchor(Point(anchorX, anchorY)) {}
+    IObject::IObject(double x, double y, double w, double h, double anchorX, double anchorY) : Position(Point(x, y)), Size(Point(w, h)), Anchor(Point(anchorX, anchorY)) {}
     std::list<std::pair<bool, IObject *>>::iterator IObject::GetObjectIterator() const {
         return objectIterator;
     }
     void IObject::Draw() const {}
-    void IObject::Update(float deltaTime) {}
+    void IObject::Update(double deltaTime) {}
 }
