+++ Engine/IObject.hpp
@@ -31,7 +31,7 @@ namespace Engine {
         /// <param name="h">Height of the image, 0 indicates original size.</param>
         /// <param name="anchorX">The centerX of the object. (0, 0) means top-left, while (1, 0) means top-right.</param>
         /// <param name="anchorY">The centerY of the object. (0, 1) means bottom-left, while (1, 1) means bottom-right.</param>
-        explicit IObject(float x, float y, float w = 0, float h = 0, float anchorX = 0, float anchorY = 0);
+        explicit IObject(double x, double y, double w = 0, double h = 0, double anchorX = 0, double anchorY = 0);
 
     public:
         // Determines whether this object should be drawn and updated.
@@ -69,7 +69,7 @@ namespace Engine {
         /// This is called when the game should update its logic, usually 'fps' times per second.
         /// </summary>
         /// <param name="deltaTime">Time elapsed since last update, can be used to calculate value changes.</param>
-        virtual void Update(float deltaTime);
+        virtual void Update(double deltaTime);
     };
 }
 #endif   // IOBJECT_HPP
