+++ UI/Component/Image.cpp
@@ -7,7 +7,7 @@
 #include "Image.hpp"
 
 namespace Engine {
-    Image::Image(std::string img, float x, float y, float w, float h, float anchorX, float anchorY) : IObject(x, y, w, h, anchorX, anchorY) {
+    Image::Image(std::string img, double x, double y, double w, double h, double anchorX, double anchorY) : IObject(x, y, w, h, anchorX, anchorY) {
         if (Size.x == 0 && Size.y == 0) {
             bmp = Resources::GetInstance().GetBitmap(img);
             Size.x = GetBitmapWidth();
