+++ UI/Component/Slider.cpp
@@ -4,7 +4,7 @@
 #include "Engine/Point.hpp"
 #include "Slider.hpp"
 
-Slider::Slider(float x, float y, float w, float h)
+Slider::Slider(double x, double y, double w, double h)
   : ImageButton("stage-select/slider.png", "stage-select/slider-blue.png", x, y),
   Bar("stage-select/bar.png", x, y, w, h),
   End1("stage-select/end.png", x, y + h / 2, 0, 0, 0.5, 0.5),
@@ -19,10 +19,10 @@ void Slider::Draw() const {
   End2.Draw();
   ImageButton::Draw();
 }
-void Slider::SetOnValueChangedCallback(std::function<void(float value)> onValueChangedCallback) {
+void Slider::SetOnValueChangedCallback(std::function<void(double value)> onValueChangedCallback) {
   OnValueChangedCallback = onValueChangedCallback;
 }
-void Slider::SetValue(float value) {
+void Slider::SetValue(double value) {
   ImageButton::Position.x=Bar.Position.x+value*Bar.Size.x;
   OnValueChangedCallback(value);
 
@@ -38,8 +38,8 @@ void Slider::OnMouseMove(int mx, int my) {
   ImageButton::OnMouseMove(mx, my);
   if (Down) {
     // Clamp
-    float clamped = std::min(std::max(static_cast<float>(mx), Bar.Position.x), Bar.Position.x + Bar.Size.x);
-    float value = (clamped - Bar.Position.x) / Bar.Size.x * 1.0f;
+    double clamped = std::min(std::max(static_cast<double>(mx), Bar.Position.x), Bar.Position.x + Bar.Size.x);
+    double value = (clamped - Bar.Position.x) / Bar.Size.x * 1.0f;
     SetValue(value);
   }
 }
